Get identifiers from other databases for a set of ott IDs

Retrieve identifiers from other databases through onezoom for a set of Open Tree of Life IDs.

Usage

ott2identifiers(otts, remove_empty = FALSE, rate = 5, key = NA, basereq = NA)

Arguments

otts

a numeric ID or numeric vector of ott ids indicating the particular taxa to evaluate.

remove_empty

remove columns if all entries are NA.

rate

maximum number of calls to the API per second.

key

your OneZoom API key (defaults to the public key of 0 if not specified).

basereq

an httr2 request object, as generated by oz_basereq(). If NA, uses the default request.

Value

A data frame of the input otts alongside their ids in other databases

Identifiers

The following identifiers are potentially returned

ncbi

ifung

worms

irmng

gbif

ipni

eol

wikidata

iucn (only for leaf node taxa)

Author

Francis Windram

Examples

if (FALSE) { # \dontrun{
ott2identifiers(247341)

ott2identifiers(c(247341, 563159, 269666, 504327, 668392), rate=5)

ott2identifiers(c(247341, 563159, 269666, 504327, 668392), remove_empty=TRUE)
} # }